fix: skip cleaning errors and array variable downsizing
In a previous change we change the store behaviour to throw an error (so we can catch it to show the expression as label) instead of just returning null. These error were not handled when resizing / cleaning and getting data. I moved try / catch up to handle the errors (they are just skipped).